requirements.txt
fastapi[standard]
httpx
python-jose[cryptography]
pydantic-settings
